+++ test/es/noisetable_test.py
@@ -10,10 +10,10 @@ def test_create_shared_arr(comm):
     shared_arr = create_shared_arr(comm, size)
 
     if comm.rank == 0:
-        shared_arr[:size] = np.arange(size, dtype=np.float64)
+        shared_arr[:size] = np.arange(size, dtype=np.float32)
 
     comm.barrier()
-    assert (shared_arr == np.arange(size, dtype=np.float64)).all()
+    assert (shared_arr == np.arange(size, dtype=np.float32)).all()
 
 
 def test_create_shared(comm):
@@ -23,4 +23,4 @@ def test_create_shared(comm):
 
     all_noise = comm.alltoall([noise] * comm.size)
     assert np.isclose(all_noise, all_noise[0]).all()
-    assert np.isclose(all_noise[0], np.random.RandomState(seed).randn(size)).all()
+    assert np.isclose(all_noise[0], np.random.RandomState(seed).randn(size).astype(np.float32)).all()
